Configuring MKL and LAPACK
Changed the configuration such that the linkage with MKL implies in the use of the Lapack version of the code.
Also corrected small interfaces that caused compiler errors in Visual Studio

+++ Mesh/pzcmesh.h
@@ -162,7 +162,7 @@ public:
 	long NElements() const {return fElementVec.NElements();}
 	
 	/** @brief Number of materials */
-	int NMaterials() const {return fMaterialVec.size();}
+	size_t NMaterials() const {return fMaterialVec.size();}
 	
 	//@}
